To unmap from in-container to on-host, use _map_targ_src
The unmap functions used the _map_src_targ function with reversed
arguments.  However, that caused a check whether the target side is a
directory on the host, which is irrelevant and can cause errors.  A
new function _map_targ_src still does the directory check, but only on
the host side.